N2N
Noise to Nouveau explores the intersection of artificial intelligence and creativity, challenging our understanding of what constitutes ‘art.’ The exhibition delves into the concept of recipes for art and the idea of varying degrees of reality in art, highlighting the dynamic collaboration between artists and AI, blurring the lines between author and tool. Visitors are invited to reflect on how the artists’ level of input influences the art-making process, questioning the boundaries of effort, authorship, and originality.